don't use "using namespace std" and use auto when possible
It's a very bad practice to use "using namespace std;" in header files,
and for consistency it's idiomatic to not put it in cpp files either.
It's bad to put it in headers because any file that includes the
header will also inherit the "using", so it's infectious. And it
can result in identifier ambiguities. See https://stackoverflow.com/a/1453605

And use auto with an initializer everywhere you can. Doing
"type name = value;" can cause implicit conversions. Implicit
conversions are BAD. Use "auto name = value;" or "auto name = type{value};"
if you want to be explicit about the type. Prefer braces over parenthesis
when calling constructors to avoid implicit conversions.

Subtle change, take const ref in findCustomTagOccurrence to avoid copying
the demo.html which might be big. But if you do, be careful not to modify
the mainHTML because the match results (smatch) holds references to
the mainHTML. In this case I think it's fine because sub_match::str()
appears to copy the string, so we extract the match before doing
the replace later.