reduce side effects and use RAII
previously you created an ifstream in one function, returned it,
and consumed it in a different function. Because fstreams are
side-effects, you have 2 untestable functions, and the lifetime
of the ifstream must be known by both functions and the outer caller.

Instead, I recommend doing all the side effects in one function readFile()
and do the path computation (no side-effects, testable) in another
function. This way, the lifetime of the ifstream is contained in one
function and you don't need to explicitly close the file because
the destructor will do it. See RAII: https://en.cppreference.com/w/cpp/language/raii